codebert-base.commons-lang.mid-2298.idx-19121.2.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableInt.java
@@ -238,7 +238,7 @@ public class MutableInt extends Number implements Comparable<MutableInt>, Mutabl
      * @since 3.5
      */
     public int addAndGet(final Number operand) {
-        this.value += operand.intValue();
+        this.value += operand.doubleValue();
         return value;
     }
